Review 연관관계 매핑 작업
Review <- N:1 -> Store
Review <- N:1 -> User

-Dto 파악하기 쉽게 기능마다 추가.
-Exception 수정 및 추가

+++ src/main/java/com/example/jmcbackend/store/dto/StoreSimpleListRes.java
@@ -21,7 +21,7 @@ public class StoreSimpleListRes {
         private String storeName; //스토어 이름
         private int storeLikeCount; //좋아요 개수
         private int storeReviewCount; //리뷰 개수
-        private double reviewAvg;        //리뷰 점수
+        private Float reviewAvg;        //리뷰 점수
         private String thumbnailImg; // or store logo
         private Long categoryId; // 음식점 분류 카테고리
         private String contactNumber; //컨택트 번호
@@ -34,7 +34,7 @@ public class StoreSimpleListRes {
         private String introduction; // 간단 가게 소개
 
         //작업중
-        public StoreSimpleListRes fromEntity(Store store,long reviewCount,double reviewAvg,long likeCount){
+        public StoreSimpleListRes fromEntity(Store store,long reviewCount,Float reviewAvg,long likeCount){
                 return  StoreSimpleListRes.builder()
                         .storeId(store.getStoreId())
                         .storeName(store.getStoreName())
